refactor: move agent element data types to modules/contexts
